feat: api folder restructure
