--- webex_assistant_skills_sdk/api/directives/__init__.py
@@ -1,22 +1,0 @@
-from webex_assistant_skills_sdk.api.directives.asr_hint import ASRHint
-from webex_assistant_skills_sdk.api.directives.assistant_event import AssistantEvent
-from webex_assistant_skills_sdk.api.directives.clear_web_view import ClearWebView
-from webex_assistant_skills_sdk.api.directives.display_web_view import DisplayWebView
-from webex_assistant_skills_sdk.api.directives.listen import Listen
-from webex_assistant_skills_sdk.api.directives.reply import Reply
-from webex_assistant_skills_sdk.api.directives.sleep import Sleep
-from webex_assistant_skills_sdk.api.directives.speak import Speak
-from webex_assistant_skills_sdk.api.directives.ui_hint import UIHint
-
-
-__all__ = [
-    'ASRHint',
-    'AssistantEvent',
-    'ClearWebView',
-    'DisplayWebView',
-    'Listen',
-    'Reply',
-    'Sleep',
-    'Speak',
-    'UIHint',
-]
