+++ webex_assistant_skills_sdk/api/shared/models/directives/speak.py
