+++ webex_assistant_skills_sdk/api/shared/services/crypto_verification.py
@@ -1,0 +1,5 @@
+from webex_assistant_skills_sdk.shared.services import CryptoService
+
+
+class CryptoVerificationService(CryptoService):
+    pass
